chore(test-fixtures): inline triplicated CapturingHandler into per-module *ConsoleSinkTest
The same 30-line `java.util.logging.Handler` lived verbatim in three places — one per format module — each with exactly one consumer (the sibling *ConsoleSinkTest). Inline as a private nested class inside the sole consumer, matching how KPipeBytesConsoleSinkTest already structures its (slightly different) handler. A shared kpipe-test-fixtures module would be over-engineering for a 30-line JUL handler and would add a new published artifact + JPMS split-package coordination for no real reuse benefit.
