+++ lib/kpipe-format-json/src/test/java/io/github/eschizoid/kpipe/format/json/JsonConsoleSinkTest.java
@@ -2,7 +2,10 @@ package io.github.eschizoid.kpipe.format.json;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
+import java.util.logging.Handler;
+import java.util.logging.LogRecord;
 import java.util.logging.Logger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -106,5 +109,31 @@ class JsonConsoleSinkTest {
     sink.accept("my-value");
     final var out = output();
     assertTrue(out.contains("\"processedMessage\":\"my-value\""), "Expected processedMessage value");
+  }
+
+  /// JUL [Handler] that appends each [LogRecord]'s raw message to an in-memory buffer.
+  /// No formatting, no parameter substitution, no separator — assertions just look for
+  /// substrings in the concatenated output.
+  private static final class CapturingHandler extends Handler {
+
+    private final ByteArrayOutputStream buffer = new ByteArrayOutputStream();
+
+    @Override
+    public void publish(final LogRecord record) {
+      if (record.getMessage() != null) {
+        buffer.writeBytes(record.getMessage().getBytes(StandardCharsets.UTF_8));
+      }
+    }
+
+    @Override
+    public void flush() {}
+
+    @Override
+    public void close() {}
+
+    @Override
+    public String toString() {
+      return buffer.toString(StandardCharsets.UTF_8);
+    }
   }
 }
